Created a setup for react navigation 5.0
