+++ package.json
@@ -32,7 +32,12 @@
     "@react-native-community/cli-platform-android": "^3.0.3",
     "@react-native-community/cli-platform-ios": "^3.0.0",
     "@react-native-community/geolocation": "^2.0.2",
+    "@react-native-community/masked-view": "^0.1.6",
     "@react-native-community/viewpager": "^3.3.0",
+    "@react-navigation/bottom-tabs": "^5.0.5",
+    "@react-navigation/drawer": "^5.0.5",
+    "@react-navigation/native": "^5.0.5",
+    "@react-navigation/stack": "^5.0.5",
     "autosuggest-highlight": "^3.1.1",
     "axios": "^0.18.1",
     "bugsnag-react-native": "^2.23.2",
@@ -74,7 +79,7 @@
     "react-native-document-picker": "^3.2.4",
     "react-native-extended-stylesheet": "^0.12.0",
     "react-native-fs": "2.14.1",
-    "react-native-gesture-handler": "^1.5.2",
+    "react-native-gesture-handler": "^1.6.0",
     "react-native-google-places-autocomplete": "^1.4.0",
     "react-native-htmlview": "^0.15.0",
     "react-native-i18n": "^2.0.15",
@@ -94,8 +99,9 @@
     "react-native-popover-tooltip": "^1.1.4",
     "react-native-ratings": "^6.3.1",
     "react-native-raw-bottom-sheet": "^2.0.6",
-    "react-native-reanimated": "^1.4.0",
+    "react-native-reanimated": "^1.7.0",
     "react-native-root-toast": "^3.2.0",
+    "react-native-safe-area-context": "^0.7.3",
     "react-native-screens": "^1.0.0-alpha.23",
     "react-native-simple-radio-button": "^2.7.4",
     "react-native-slick": "^1.0.2",
@@ -105,7 +111,7 @@
     "react-native-tab-view": "^2.11.0",
     "react-native-vector-icons": "6.6.0",
     "react-native-webview": "^7.5.2",
-    "react-navigation": "^4.0.10",
+    "react-navigation": "^4.1.1",
     "react-navigation-drawer": "^2.3.3",
     "react-navigation-redux-helpers": "^1.1.2",
     "react-navigation-stack": "^1.10.3",
